+++ src/Jarboe/Table/Fields/Checkbox.php
@@ -3,6 +3,7 @@
 namespace Yaro\Jarboe\Table\Fields;
 
 use Illuminate\Http\Request;
+use Yaro\Jarboe\Table\Fields\Traits\Inline;
 use Yaro\Jarboe\Table\Fields\Traits\Nullable;
 use Yaro\Jarboe\Table\Fields\Traits\Orderable;
 
@@ -10,6 +11,7 @@ class Checkbox extends AbstractField
 {
     use Orderable;
     use Nullable;
+    use Inline;
 
     public function value(Request $request)
     {